+++ database/migrations/2020_11_09_081900_create_dealer_product_histories_table.php
@@ -16,7 +16,7 @@ class CreateDealerProductHistoriesTable extends Migration
         Schema::create('dealer_product_histories', function (Blueprint $table) {
             $table->id();
             $table->integer('dealer_product_id');
-            $table->tinyInteger('type');
+            $table->integer('type');
             $table->integer('quantity');
             $table->text('detail');
             $table->integer('user_id');
